+++ src/db/versions.idb.ts
@@ -9,4 +9,13 @@ export function defineVersions(db: Dexie7) {
 		entitlements: "id,scope,timestamp,user_id",
 		settings: "key",
 	});
+
+	db.version(2.5).stores({
+		channels: "id,timestamp",
+		emoteSets: "id,timestamp,priority,provider,scope",
+		emotes: "id,timestamp,name,owner.id",
+		cosmetics: "id,timestamp,kind",
+		entitlements: "id,scope,timestamp,user_id",
+		settings: "key",
+	});
 }