+++ src/types/module.d.ts
@@ -1,3 +1,4 @@
+import type AvatarsModuleVue from "@/site/twitch.tv/modules/avatars/AvatarsModule.vue";
 import type ChatInputModuleVue from "@/site/twitch.tv/modules/chat-input/ChatInputModule.vue";
 import type ChatModuleVue from "@/site/twitch.tv/modules/chat/ChatModule.vue";
 import type EmoteMenuModuleVue from "@/site/twitch.tv/modules/emote-menu/EmoteMenuModule.vue";
@@ -6,8 +7,9 @@ import type SettingsModuleVue from "@/site/twitch.tv/modules/settings/SettingsMo
 declare type ModuleID = keyof ModuleComponentMap;
 
 declare type ModuleComponentMap = {
-	chat: typeof ChatModuleVue;
 	"chat-input": typeof ChatInputModuleVue;
-	settings: typeof SettingsModuleVue;
 	"emote-menu": typeof EmoteMenuModuleVue;
+	avatars: typeof AvatarsModuleVue;
+	chat: typeof ChatModuleVue;
+	settings: typeof SettingsModuleVue;
 };